+++ packages/react-diagrams-routing/src/link/Names.ts
@@ -1,0 +1,2 @@
+export const PATH_FINDING_LINK_FACTORY_NAME = 'pathfinding';
+export const RIGHT_ANGE_LINK_FACTORY_NAME = 'rightAngle';
