+++ packages/react-diagrams-routing/src/link/PathFindingLinkModel.ts
@@ -1,10 +1,10 @@
-import { PathFindingLinkFactory } from './PathFindingLinkFactory';
 import { DefaultLinkModel, DefaultLinkModelOptions } from '@projectstorm/react-diagrams-defaults';
+import { PATH_FINDING_LINK_FACTORY_NAME } from './Names';
 
 export class PathFindingLinkModel extends DefaultLinkModel {
 	constructor(options: DefaultLinkModelOptions = {}) {
 		super({
-			type: PathFindingLinkFactory.NAME,
+			type: PATH_FINDING_LINK_FACTORY_NAME,
 			...options
 		});
 	}
