+++ packages/react-diagrams-routing/src/link/RightAngleLinkWidget.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { DiagramEngine, LinkWidget, PointModel } from '@projectstorm/react-diagrams-core';
 import { RightAngleLinkFactory } from './RightAngleLinkFactory';
-import { DefaultLinkModel, DefaultLinkSegmentWidget } from '@projectstorm/react-diagrams-defaults';
+import { DefaultLinkSegmentWidget } from '@projectstorm/react-diagrams-defaults';
 import { Point } from '@projectstorm/geometry';
 import { MouseEvent } from 'react';
 import { RightAngleLinkModel } from './RightAngleLinkModel';
